pyproject.toml
[project]
name = "osapiens-challenge-makeathon-2026"
version = "0.1.0"
description = "Materials for the osapiens Makeathon 2026 deforestation detection challenge."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "geopandas~=1.0.0",
    "numpy~=1.26.4",
    "matplotlib~=3.9.0",
    "pandas~=2.2.2",
    "rasterio~=1.4.3",
    "shapely~=2.0.3",
    "boto3~=1.34",
    "botocore",
    "tqdm~=4.66",
    "ipykernel",
    "jupyter",
]

[dependency-groups]
dev = [
    "black>=24.4.0,<25.0.0",
    "coverage>=7.6.3,<8.0.0",
    "flake8>=7.0.0,<8.0.0",
    "flake8-pyproject>=1.2.3,<2.0.0",
    "isort>=5.13.2,<6.0.0",
    "mypy>=1.9.0,<2.0.0",
    "pre-commit>=3.7.0,<4.0.0",
    "pre-commit-hooks>=4.6.0,<5.0.0",
    "pytest>=8.3.2,<9.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
]
experiments = [
    "catboost>=1.2.10",
    "imbalanced-learn>=0.14.1",
    "lightgbm>=4.6.0",
    "optuna>=4.8.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "shap>=0.49.1",
    "shinka-evolve>=0.0.5",
    "xgboost>=3.2.0",
]
gpu-experiments = [
    "accelerate>=1.13.0",
    "albumentations>=2.0.8",
    "einops>=0.8.2",
    "kornia>=0.8.2",
    "lightning>=2.6.1",
    "pytorch-triton-rocm==3.5.1 ; python_version < '3.14' and sys_platform == 'linux' and platform_machine == 'x86_64'",
    "safetensors>=0.7.0",
    "segmentation-models-pytorch>=0.5.0",
    "timm>=1.0.26",
    "torch>=2.9.1 ; python_version < '3.14' and sys_platform == 'linux' and platform_machine == 'x86_64'",
    "torchaudio>=2.9.1 ; python_version < '3.14' and sys_platform == 'linux' and platform_machine == 'x86_64'",
    "torchgeo>=0.6.2",
    "torchmetrics>=1.9.0",
    "torchvision>=0.24.1 ; python_version < '3.14' and sys_platform == 'linux' and platform_machine == 'x86_64'",
]

[tool.uv]
package = false
default-groups = ["dev", "experiments"]

[tool.uv.sources]
torch = { index = "pytorch-rocm" }
torchaudio = { index = "pytorch-rocm" }
torchvision = { index = "pytorch-rocm" }
pytorch-triton-rocm = { index = "pytorch-rocm" }

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.3"
explicit = true

[tool.black]
line-length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
    ".",
]

[tool.coverage.run]
source = [
    "download_data",
    "submission_utils",
]
omit = [
    "tests/*",
]

[tool.coverage.report]
skip_empty = true
skip_covered = true
